backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
async-timeout==5.0.1
billiard==4.2.2
Brotli==1.1.0
cachetools==6.2.1
celery==5.3.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.8.0
distro==1.9.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.0
dnspython==2.8.0
fonttools==4.60.1
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud==0.34.0
googleapis-common-protos==1.70.0
groq==0.32.0
grpc-stubs==1.53.0.6
grpcio==1.75.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
kombu==5.5.4
packaging==25.0
pillow==11.3.0
pdf2image==1.17.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pydyf==0.11.0
pymongo==4.15.3
pyparsing==3.2.5
pypdf==3.17.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==4.4.4
reportlab==4.4.3
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tinycss2==1.4.0
types-protobuf==4.25.0.20240417
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==52.5
webencodings==0.5.1
zopfli==0.2.3.post1

python-dotenv==1.1.1

# Stripe Integration for Payments
stripe==8.9.0

# Exchange Rate API for Multi-Currency
requests
